Replace UEBERZUG with Viu.
Viu is a bit different and works a bit different than ueberzug, but it
seems to do a nice job.

I managed to get everything working on my Macbook too (i'll include a
movie in the pull request)

Some limitations:
I didn't really know how that FIFO system worked, but I figure it was
used to not make the terminal preview window go haywire on reloads
(drawing lines and clearing the screen). So I have left GIFs unanimated
(it is an out-of-the-box feature of VIU though)

On Linux (given that you aren't running inside tmux the images will look
nice and crisp, in tmux and on OSX the images are instead rendered to
block-maps.. looks a bit worse, but still fairly usable)

Note one:
It's worth mentioning that the previous version (with ueberzug) used
imagemagick in gifs to find frames and convert frames to images.
I thought about adding docs for it, but I'm not generating stills
through scripting anymore. Make sure to add the requirement to
imagemagick to the docs if generating stills becomes a thing again ;)

Note two:
the next upcoming verison of viu allows x & y offset, but that release
is still in the works. So i haven't included it in this pull request. I
might create a new pull request to re-add the offset in a few weeks.
Until then I think it's okay to keep that info around when we call the
vimg script.

Thanks for looking at this pull and thanks for the amazing work on this
plugin!